+++ observability/akka-http-metrics/src/test/suite/scala/com/daml/metrics/akkahttp/OpenTelemetryUtils.scala
@@ -3,8 +3,6 @@
 
 package com.daml.metrics.akkahttp
 
-import scala.annotation.tailrec
-
 import com.daml.metrics.api.MetricHandle
 import com.daml.metrics.api.opentelemetry.OpenTelemetryFactory
 import io.opentelemetry.api.metrics.{Meter => OtelMeter}
@@ -14,70 +12,6 @@ import io.opentelemetry.sdk.testing.exporter.InMemoryMetricReader
 import io.opentelemetry.sdk.testing.time.TestClock
 import java.time.Duration
 import java.util.concurrent.atomic.AtomicInteger
-
-/** Data structure used to compare the content of histogram metrics
-  * The exact recorded values are not accessible, using the other attributes to check that
-  * the data was correctly recorded.
-  */
-case class HistogramData(
-    count: Long,
-    sum: Long,
-    bucketCounts: List[Long],
-)
-
-object HistogramData {
-  final val defaultBucketBoundaries =
-    List(5L, 10L, 25L, 50L, 75L, 100L, 250L, 500L, 750L, 1000L, 2500L, 5000L, 7500L, 10000L)
-  final val bucketCountsZero = List.fill(15)(0L)
-
-  /** Generates a HistogramData instance, from the given value, to compare with the data
-    * extracted from a metric.
-    */
-  def apply(value: Long): HistogramData = {
-    HistogramData(List(value))
-  }
-
-  /** Generates an HistogramData instance, from the given values, to compare with the data
-    * extracted from a metric.
-    */
-  def apply(values: List[Long]): HistogramData = {
-    computeData(values, HistogramData(0, 0, bucketCountsZero))
-  }
-
-  @tailrec private def computeData(values: List[Long], acc: HistogramData): HistogramData = {
-    values match {
-      case head :: tail =>
-        computeData(
-          tail,
-          HistogramData(
-            acc.count + 1,
-            acc.sum + head,
-            updateBucketCounts(acc.bucketCounts, defaultBucketBoundaries, head),
-          ),
-        )
-      case Nil =>
-        acc
-    }
-  }
-
-  private def updateBucketCounts(
-      counts: List[Long],
-      boundaries: List[Long],
-      value: Long,
-  ): List[Long] = {
-    boundaries match {
-      case head :: tail =>
-        if (value <= head) {
-          (counts.head + 1) :: counts.tail
-        } else {
-          (counts.head) :: updateBucketCounts(counts.tail, tail, value)
-        }
-      case Nil =>
-        // last bucket
-        List(counts.head + 1)
-    }
-  }
-}
 
 /** Base class to manage metrics in a test.
   * Use the metric factory to create the metrics.
